Refactor functions inside different files
